* Backport compiler fixes from 5.1   * Fix warnings with Win64.   * Fix IP/PGM Router Alert with Win64.   * Add further restrict usage.   * Remove unused pthread_rwlock_t from Windows slim reader/writer lock usage.   * Update SNMP & HTTP modules for new peer/socket usage.   * Add error checking for non-Win32, non-GNU compilers in thread module.   * Fix select usage in recovery of hard failure on network sends on POSIX platforms. * Bump micro version.

+++ branches/release-5-0/openpgm/pgm/wsastrerror.c
@@ -357,7 +357,7 @@ pgm_win_strerror (
 	const int	win_errno
 	)
 {
-	const DWORD nSize = buflen;
+	const DWORD nSize = (DWORD)buflen;
 	FormatMessage (FORMAT_MESSAGE_FROM_SYSTEM,
 		       NULL,		/* source */
 		       win_errno,	/* message id */
